requirements.txt
Flask==1.0.2
uWSGI==2.0.17.1
virtualenv==16.0.0
paho-mqtt==1.4.0